[ANT] Add ability to toggle play/pause
Added functionality to toggle play or pause. When
the board receives a byte corresponding to the
play/pause message, it will keep track of it as
a "sequence number". If this byte changes, then it
will toggle play or pause.

Tested by sending various bytes from ANTWare and
saw that play/pause is being toggled. Also tested
with the manual button pressing method.