[CIQ] Open ANT channel on select button press
Open a master ANT channel when the select button is pressed.
The ANT channel currently just opens and does nothing else.
At this point, there is no way to close the channel (besides
closing the app), and there is no protection to opening more
than 8 channels.

Tested on simulator that the onMessage() function callback is
displaying messages when the channel is open.