[CIQ] Add placeholder "connecting" view
Added a placeholder view which is meant to show
a "connecting" message when the ANT channel is
being open. For now, it will have placeholder text.

Functionality currently remains the same - Pressing
the select button will take you to this placeholder
page. Pressing next page button will open the menu
to send messages.