[MIDI] Add script to generate code from tracks
Added a script that takes in up to 2 MIDI files with
single tracks, and generates code for the firmware.
Notes are translated into frequencies and delta times
are translated into milliseconds.

The current output of the script gives arrays of notes
and should be able to easily copy and paste into the
current MIDI player prototype.