[CIQ] New view after ANT connection view
Added a new view that follows the ANT connection view
after the channel is formed. This view acts as the main
U.I. for selecting music controls. This view acts like a
menu, where pressing up and down buttons will scroll
through options. As of right now, the menu options are just
text, to be replaced with icons.

Also added a class that acts as a counter between objects.
This will keep track of which menu item to show on the current
view.

Verified scrolling between items in the new view acts like
a menu. Pressing select sends an ANT message as before. Back
button takes you back to the main page and release the channel.