[Music] Save code space on duplicate buzzer arrays
Re-generated notes code for songs that play the same
track on each buzzer. Then, manually changed it so
that each buzzer points to the same notes array to
reduce code space by about half for that song.

Also added max. uint16 protection in the MIDI code
generator script to prevent compiler warnings.
Typically occurs when the script is trying to pad
a zero note for the duration of the whole song.