[ANT] Disable sending ACK message
This is an attempted fix to prevent the EiE board
from completely locking up after a period of time
while the CIQ app. is connected. Still not sure the
reason why it does that, but disabling ACK's appear
to help.

As a result, used a #define statement in the firmware
to "comment" out sending of ACK messages. In CIQ, the
ANT connection view will now simply transition to the
music controls after a 3 seconds delay.