[CIQ] Detect forward and backward tilt
From the "raised static" state, added logic to detect
if the watch is being tilted forward or backwards
(relative to having the watch on the left hand).

Also modified debug counters on the view. Left number
shows which state it is in, middle shows number of tilt
forward detections, and right shows number of tilt
back detections