[CIQ] Show instructions in gesture control view
Remove sensor debug items and show instructions.
If device is not yet still, it will show a message
saying to hold the device flat and still. When it is
in that state, the message will say to tilt the watch
to change songs.

Also, changed state name from "raised static" (what
does this even mean...?) to "hold flat".